refactor : notification폴더 global 안으로 이동, 레포지토리 이름 변경
