+++ server2/server-common/src/main/java/sinhan/server2/global/notification/mongo/NotificationRepository.java
@@ -1,11 +1,11 @@
-package sinhan.server2.notification.mongo;
+package sinhan.server2.global.notification.mongo;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
-import sinhan.server2.notification.entity.Notification;
+import sinhan.server2.global.notification.entity.Notification;
 
 import java.util.List;
 
-public interface newNotificationRepository extends MongoRepository<Notification, Integer> {
+public interface NotificationRepository extends MongoRepository<Notification, Integer> {
     public void deleteByNotificationSerialNumber(String notificationSerialNumber);
 
     public void deleteByReceiverSerialNumber(Long receiverSerialNumber);
